+++ lib/page/book_player/epub_player.dart
@@ -192,6 +192,7 @@ class EpubPlayerState extends ConsumerState<EpubPlayer>
         customCSS: `${Prefs().customCSS.replaceAll('`', '\\`')}`,
         customCSSEnabled: ${Prefs().customCSSEnabled},
         useBookStyles: ${Prefs().useBookStyles},
+        headingFontSize: ${style.headingFontSize},
       })
       ''');
     });